requirements.txt
langchain==0.0.178
openai==0.27.7
matplotlib==3.7.1
ipython==8.13.2
termcolor==2.3.0
Pillow==9.5.0
